requirements.txt
# Core Streamlit
streamlit>=1.28.0
langchain-google-vertexai
transformers
langchain_qdrant
langchain
langchain-community
torch
numpy
pdfplumber
python-dotenv
requests
sentence_transformers
langgraph
langchain-openai
